Change id field to auto inc.

+++ database/migrations/2022_07_21_061542_create_entity_types_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('entity_types', function (Blueprint $table) {
-            $table->unsignedBigInteger('id');
+            $table->increments('id');
             $table->string('label');
             $table->timestamps();
         });
